requirements.txt
json-rpc
azure-mgmt-redis